fix: move drain timeout error outside the store boundary
The kernel index may not re-export store internals; relocate
DaemonDrainTimeoutError and the queue drain target shape to
kernel/daemon so the public export respects the import boundary.
