+++ packages/kernel/src/index.ts
@@ -149,7 +149,7 @@ export {
   readContentAddressedTextBlob,
   writeContentAddressedBlob
 } from "./composition/index.ts";
-export { DaemonDrainTimeoutError } from "./store/daemon-drain.ts";
+export { DaemonDrainTimeoutError } from "./daemon/drain-timeout.ts";
 export { writeCoordinatedPayload, writeCoordinatedTaskDocuments } from "./write-coordination/write-helpers.ts";
 export {
   readDaemonRegistry,
