+++ packages/kernel/src/store/daemon-drain.ts
@@ -1,14 +1,5 @@
-import type { DaemonQueueDrainTarget, DaemonWriteQueue } from "./daemon-runtime-queue.ts";
-
-export class DaemonDrainTimeoutError extends Error {
-  readonly targets: ReadonlyArray<DaemonQueueDrainTarget>;
-
-  constructor(rootDir: string, drainTimeoutMs: number, targets: ReadonlyArray<DaemonQueueDrainTarget>) {
-    super(`daemon queue drain timed out after ${drainTimeoutMs}ms for ${rootDir}: ${targets.map(describeDrainTarget).join(", ") || "unknown in-flight operation"}`);
-    this.name = "DaemonDrainTimeoutError";
-    this.targets = targets;
-  }
-}
+import { DaemonDrainTimeoutError } from "../daemon/drain-timeout.ts";
+import type { DaemonWriteQueue } from "./daemon-runtime-queue.ts";
 
 export async function waitForDaemonQueueIdle(
   queue: DaemonWriteQueue,
@@ -27,10 +18,4 @@ export async function waitForDaemonQueueIdle(
   } finally {
     if (timeout) clearTimeout(timeout);
   }
-}
-
-function describeDrainTarget(target: DaemonQueueDrainTarget): string {
-  return target.kind === "interactive"
-    ? `interactive command ${target.commandId} (${target.opIds.join(",")})`
-    : `background source ${target.source}`;
 }
